Added mean_and_nines/2 to use Dizzy's stats lib
The stats lib method is 2x faster than using file_sorter method and
similar or less amount of RAM and far, far less than slide.erl's
original RAM-only implementation.  It is quite a bit slower than the
RAM-only implementation for actually calculating the percentile stats,
but the purpose of this exercise was reducing overall memory required
to maintain the percentile stats for slide and reduce the amount of
CPU required for slide (by omitting calculations for data that Riak
never uses).