Fix slide.erl unit tests.
The new implementation of slide cannot use only the moment "time" to
make decisions of which 1-second files to consider: time must actually
go by in 1 second increments (to allow file mtimes to change).  So,
to simulate the passage of time for EUnit tests, I added the
idle_time_passing/3 function.  A side-effect of this is adding
extra event counts, one per simulated second, so some of the expected
counts from the pre-commit tests needed to be increased.

Also, the quantile calculations by basho_stats_histogram.erl are
clever estimates, but the ranges of 1-100 needed to be increased by
a factor of 100 to be able to fall into different histogram summary bins.