Return `{Idx,Pid}` pairs from `all_nodes`
This information is useful for a vnode to determine if it has a
"predecessor."  A predecessor is the vnode that was covering the
same index before the "new" vnode was started.  This info can be
used during node leave/join to proxy request from one vnode to
another in order to avoid not_found replies.