Generalize the coverage behavior used to list keys and move much of the coverage work into riak_core.
Fixes: az496

This pull request generalizes the coverage behavior used by key
listing and makes it directly usable for other operations.

Also added function to riak_core_ring to determine the ring index that
is the successor of a chashed index value directly given the number of ring
partitions instead of having to calculate the entire preference
list. There is also a related function added to the chash module and
a quickcheck test to exercise this function.

This pull request also includes some changes to tidy up some of the
chash and riak_core_ring modules by adding some formatting, converting
edoc to specs, etc.