Fix bug with nodes leaving the cluster earlier than intended.
Change ring_ready to wait on exiting nodes in addition to valid and leaving
nodes. This ensure the ring converges on a node's intent to leave before the
node leaves the cluster.

Change claimant from moving itself from exiting to invalid. Instead, after
the claimant moves to exiting, a new claimant will emerge that will move the
previous claimant to invalid and initiate shutdown.