Add 1 & 2 arity claim APIs
The claim APIs currently require both 1 & 2 arity functions
because of the two different ways legacy gossip and new gossip
call claim.

The reason both default and v1 are exported is because soon the
default will be v2 and you still need a way to allow the user to
set the claim algo to v1.