Change unregister_vnode to pass-through vnode proxy
Change the vnode manager to use an async call to notify the vnode
proxy to unregister a vnode pid. Likewise, move the logic that
sends the 'unregistered' event to the vnode from the vnode manager
to the vnode proxy. This allows the async approach to be correct,
as well as guarantees that the 'unregistered' event is the final
message in the vnode's mailbox. As previously implemented, requests
sent through the vnode proxy could arrive at the vnode after the
vnode manager had unregistered it.