Merge of slf-core-ets-match-replacement + hand-edits onto master.
Avoid use of ets:match() in hot code paths

Profiling with DTrace on Solaris suggests that ets:match() is causing
a much higher than necessary lock contention rate within the Erlang
R14B03 and R14B04 VMs.  Eliminating ets:match() in the riak_core and
riak_kv applications is substantial when using Pthread locking instead
of Ericsson's hand-tuned locking (which is the default).  The results
are more ambiguous when the default locking strategy is used.

Hand-edits were required due to changes in master after the 1.0 branch
was created.