Change how the capability system handles unknown capabilities
Change riak_core_capability:get/1 to throw an exception if it is unable
to determine the mode for the requested capability.

Add riak_core_capability:get/2 that takes a default response as the second
argument, and returns the default when the mode is unknown.

Change riak_core_vnode_master to default to legacy routing mode when the
current routing mode is unknown.