Add replace and force-replace cluster operations
Add support for `replace node new-node` operation that stages a
request for node to handoff all of its partitions to new-node and
then leave the cluster. This is similar to a leave, but directs
handoff to a specific replacement node.

Add support for `force-replace node new-node` operation that stages
a request that reassigns all partitions owned by node to new-node
and removes node from the cluster. This is similar to a force-remove,
but reassigns partitions to a specific replacement node.