Added claim_v3 algorithm.
Claim V3 - unlike the v1/v2 algorithms, v3 treats claim as an optimization problem.
In it's current form it creates a number of possible claim plans and evaluates
them for violations, balance and diversity, choosing the 'best' plan.

Violations are a count of how many partitions owned by the same node are within target-n
of one another. Lower is better, 0 is desired if at all possible.

Balance is a measure of the number of partitions owned versus the number of partitions
wanted.  Want is supplied to the algorithm by the caller as a list of node/counts.  The
score for deviation is the RMS of the difference between what the node wanted and what it
has.  Lower is better, 0 if all wants are mets.

Diversity measures how often nodes are close to one another in the preference
list.  The more diverse (spread of distances apart), the more evenly the
responsibility for a failed node is spread across the cluster.  Diversity is
calculated by working out the count of each distance for each node pair
(currently distances are limited up to target N) and computing the RMS on that.
Lower diversity score is better, 0 if nodes are perfectly diverse.