Fix 'riak-admin join'
Fix 'riak-admin join' so that it once again works despite the changes
made to node joining as part of the new staged clustering changes. This
is accomplished through the addition of auto-joining logic in the
claimant that transitions 'joining' nodes to 'valid' nodes that were
joined to the cluster using the traditional join rather than the newer
staged join.