Don't let `max_concurrency` crash handoff manager
Check for `{error, max_concurrency}` return from `send_handoff` and
create a `#handoff_status` with status of `max_concurrency` so that
it will be retried at the next available opportunity.

This is a bit hackish but if I put this code in `send_handoff` then I
have to modify how hinted handoff works.