Make xfer async and idempotent
The synchronous xfer call from vnode mgr to handoff mgr was very
brittle.  Firstly, the handoff sender needs the vnode pid and the pid
comes from the vnode manager but if the src and target partitions are
on the same node then a deadlock will occur when the handoff mgr tries
to get the pid from the vnode mgr which is still waiting for the
handoff mgr to return, thus a deadlock.  Secondly, errors can occur
during the call or max concurrency could be reached.

Rather than code for these cases it's easier to make the xfer API
idempotent and have the vnode manager constantly retry until it has
confirmation of a completed xfer.