Check repair preconditions on owner
Make the various precondition checks for repair on the owning node of
the partition.  This is better because the repair call could be made
from any node but ideally you want to check the ring and node status
from the owning node for the simple fact that nodes don't always agree
on the ring or the state of other nodes.