Fix bug that caused ring to be forcibly updated continuously
Change riak_core_claimant to only force a ring update of a stalled
ring if the ring is "ready" and has therefore already converged across
the cluster. Without this change, ring convergence always appears to
be stalled if a node is offline and therefore the force update happens
over and over.