Don't crash the cache when a stat mod crashes
Since calls to calculate stats are spawn_linked to ensure
that the calculating processes exits if the cache exits, the cache
will crash if a spawned process exits. This is bad. This change
traps_exits on spawned processes and enforms awaiting callers of
the error, without crashing the cache or effecting other registered
mods.