Remove gossip tests
These tests were originally written against an older version of riak
core and I couldn't see how to modify them to work with the new code.
I figure it is better to remove this code so that the other tests can
be added and this year old PR can finally be merged.