Make vnode check for existing handoff before starting another
Change riak_core_vnode to keep track of the handoff sender pid and not
initiate a new handoff if an existing handoff is already running. In
practice, this is not necessary as the handoff manager ensures that
only one handoff is running per source/target pair. This change is an
optimization to avoid additional work that is potentially expensive on
some backends.