Change vnode manager API to read from ETS when possible
Various parts of Riak regularly call the all_vnodes/0, all_vnodes/1,
and all_index_pid/1 functions provided by riak_core_vnode_manager.
Prior to this commit, all three calls performed gen_server calls to
the vnode manager. This forces unnecessary serialization and can be
bad news if the vnode manager ever ends up with a large message queue,
which happens frequently during cluster transitions.

This commit changes the existing private ETS table that the vnode manager
uses to keep track of vnodes, and makes the table protected. The three
API calls above are then changed to directly read from this ETS table
when possible, and only fallback to gen_server calls to the vnode manager
when necessary.  For example, if an API call comes in while the vnode
manager is still starting up, then the ETS table may not yet exist, and
the the calling process will fallback to a call to the vnode manager
which is guaranteed to be handled after the ETS table has been created
and populated.