ability to shrink ring
not many changes were necessary with the exception that when the ring shrinks
some indexes disappear, which needs to be handled specially. Note: this only works
assuming NewRingSize < MaxN, where MaxN is the largest N value used on any bucket
in the cluster. Also, reduce impact on gossip during resize using 'set_only'.