Implement SSL support for riak_core_connection and riak_core_service_mgr
This is a port of the SSL implementation from Riak's MDC implementation.
The app.config arguments are the same, only now they're under riak_core.

SSL is negotiated right after capabilities are exchanged, so minimal
information is sent 'in the clear'. If one side requests SSL and the
other side does not have it enabled, SSL is not allowed to connect.