support for when SHTF during ring resize
allow ring resize operation to be aborted. The operation is staged like
any other cluster op. The abort will only be staged/performed if the
operation is in the process of performing resize transfers (resized ring
has not been installed). Otherwise, there is nothing to abort, so
the operation is ignored.

Support force-replace during resize. Works like force-replace during
typical cluster operations. Transfers are rescheduled on the replacing
node under the assumption that the replacing node may have been rebuilt
from a recent backup.

force-remove is not supported because it may end up causing too
many transfers to be rescheduled on the remaining nodes, overloading
them. Instead force-remove is supported by performing an abort
of the resize operation first.