Change vnode + capability managers to use intelligent ring polling
The standard ring event handler sends an event for every ring received.
The vnode manager and capability manager do not care about every ring
received, they only care if the local ring actually changes. In other
words, if a ring is received that provides no new information, and
therefore the local ring does not change, there is no reason to trigger
an event.

This commit changes these managers to no longer listen to ring events,
but instead poll the ring manager at a fixed frequency. This not only
solves the above problem, but also allows the managers to ignore high
frequency ring events.

Given that the common case is for the ring to remain stable, this change
also introduces the notion of the "ring id" which is a monotonic sequence
number for the ring maintained by the ring manager. The periodic polling
then simplifies to checking if the ring id has changed, and only
performing the more expensive state update if so.