cleanup after resized ring
After a ring resizes there are several cleanup operations that must
be performed. In some cases a partition no longer exists (if the ring
shrinks) or its owner has been moved to another node. In these cases the
vnode must be unregistered and all its data deleted. In other cases, where
the vnode does continue to exist and has not moved, some data must be removed
but the vnode continues to run. This commit addresses the former.

After completion, the resized ring will have deletions scheduled for the
appropriate indexes. The deletions are scheduled like any other ring transition.
Although this is considered part of the resize operation, at this time the
future ring has been installed and the vnodes targetted for cleanup appear as fallbacks.
After each deletion is complete, the vnode unregisters. After all deletions are
complete the full resize operation is considered complete.

Also move starting of extra vnode proxies to ring handler