defense against invalid claim during resize planning
in certain cases, like when never_wants_claim is used or there are pending
changes to the ring, the claim function will not rebalance the ring. This can
cause problems in two ways. First, resizing modifies the original chash. Second,
the unclaimed future ring will have partitions owned by a dummy owner. Resizing
previously assumed the ring would always be rebalanced during claim. This is incorrect
and can lead to troublesome ownership information because of the reasons above. To
prevent against this the resized ring is now validated for any improper ownership.