only silently drop DOWN-normal messages in deleted modstate
This is a restriction of the modification made in PR #334.

Dropping all {'DOWN',_,process,_,normal} messages on the floor instead
of passing them to vnode handle_info functions causes riak_pipe vnodes
to missing messages that it uses to cleanup workers for pipes that
shutdown unexpectedly.

This commit restricts the DOWN-normal message dropping to the case
that the vnode's modstate is {deleted, _}. PR #334 suggests the original
modification was made only to quiet the log spam generated by the
following clause, which also only operates in modstate-deleted.

Before this commit, the riak_test pipe_verify_exceptions would fail
during its verify_middle_fitting_normal test, because workers would be
left running after the fitting exited 'normal'. After this commit,
workers are once again terminated correctly, so the test passes again.