Use raw ring when setting bucket props
Fix the ring trans function used by `set_bucket` so that it uses the
passed-in, raw ring.  This is important because otherwise things like
bucket prop fixups which are performed on "my ring" can leak into the
raw ring casuing issues such as a Riak Searh pre-commit hook that
exists even when the search property is set to false.