Cluster Metadata: Track Local Changes w/ hashtree_tree
Use hashtree_tree to store hashes for data in
riak_core_metadata_manager.  The tree is managed by
riak_core_metadata_hashtree. Hashes are inserted into the tree,
synchronously (for now) by metadata manager in the write path. In
addition, metadata_hashtree builds the tree with hashes for data
already stored on disk (ensuring not to trample new updates in the
process). Once the tree is built, it is periodically updated. Minus
the periodic update, the metadata_hashtree process is akin to
riak_kv_index_hashtree in Riak's AAE subsystem.

The API to access hashes is exposed via riak_core_metadata:prefix_hash/1.