Add Remote Iterators to Metadata Manager
Iterating from a node other than the one running the metadata manager
was not possible, previously -- ets continuations cannot cross
nodes. This is intedended for use in repair after hashtree exchanges.

In addition, iterator_close/1 has been added. This is necessary to cleanup
remote iterators. For local iterators it is a no-op.

This API is not exposed via riak_core_metadata since the iterface is
primarly meant to be local.