Remove riak_core_gen_server (gen_server2)
This version of gen_server2 is over 3 years old.  There have been at
least 15 comments to the canonical OTP gen_server in the last 3 years.
The longer we stay with this outdated copy of gen_server the higher
the probability of hitting a bug that is fixed in OTP.  We also miss
out on an performance benefits that might be made in OTP.  Speaking of
perf, there is no evidence that gen_server2 helps us at all.  It's
original purpose was to avoid lots of redundant work when doing
selective receive on large mailboxes.  Only the handoff receiver is
still using this behavior and it has an ack-based model to prevent
overflow.  We have already moved all other processes off of this
behavior and it has been agreed, twice over, by the majority of the
engineering team to remove gen_server2 completely.