increase broadcast's initial eager set size for larger clusters
increasing the fanout reduces the chance of requiring exchange to deliver
messages to all alive nodes during a broadcast. This patch decreases the
likelyhood by 20 to 30 percent, on average.

The chosen fanout is determined by the discrete function
`round(log(NumNodes)+1)`, where `log(x)` is the natural log and `NumNodes`
is the number of nodes in the cluster.  For practical cluster sizes,
this curve has a maximum fanout of about six peers, dependning on what
is considered the upper limit of practical cluster sizes.

addresses #438