remove legacy forwarding code that snuck in as part of 632af2b3ce9
Handling of forwarding code was changed in #301. Changes made as part
of the commit mentioned above were most likely made prior to that PRs merge.

Remove the forwarding code as it is redundant (the decision to forward
is handled further up the call stack). It caused vnode when resizing to
crash, however, since the check was not as comprehensive as its
caller.  The vnode would crash when attempting to forward to a
partition identified by `{Index, [{Index, Owner}]}` which is invalid.