Add a capability to guard riak_core_net_ticktime
The goal of this change is to prevent problems in a mixed version
cluster where the riak_core_net_ticktime module is not loaded. The
capability is net_ticktime and has two modes: supported and
unsupported. The supported mode is the preferred mode. The
net_ticktime daemon will not run until the capability has a mode of
supported.

This change also adds `riak_core_net_ticktime:enable/0` that can be
used to register the net_ticktime capability after the
riak_core_net_ticktime module is loaded on a node.