User iterator prefetching in hashtree module
Use prefetch and prefetch_stop actions for iterator_move calls in
the hashtree module. This replaces the use of the next action with the
goal of making segment traversal more efficient. This change adds a
prefetch field to the itr_state record and some additional case
clauses in the iterate/2 function. This ensures that an iterator can
be safely used with other actions after being used for prefetch
iteration.