Re-order Resize Transfers so Partitions Not Changing Ownership Go First
This commit provides a stop-gap measure to prevent a degenerate case
when resizing a ring with larger datasets. When a ring is resized, it
is possible that some partitions do not change ownership. If these
partitions have data transferred to them before being allowed to
perform their transfers they are regularly iterating over and ignoring
a bunch of data from the new ring. In some cases, this can cause
handoff to timeout because the sender fold is spending too much time
skipping items not bound for the current reciever. This can lead to
the stalling of the resize operation itself. A temporary workaround
has been to raise the handoff timeout values to let these partitions
complete their transfers, however, with this commit, these transfers
are scheduled first -- taking advantage of the next list being treated
as ordered by the vnode manager. Although this is by no means a
solution in 100% of cases it should be a general improvement and
should prevent the need for timeout adjustments, assuming when folding
that the data is well-distributed accross the various preflists the vnode
is responsible for a portion of.