Made tcpmon aware of ssl sockets for stats.
Intead of failing to get stats and then forgetting the socket completely,
the tcpmon now unwraps the socket if needed to get at the gen_tcp socket
center for stats. This seemed a better place for the unwrap than requiring
callers of the 'monitor/2' function to unwrap the socket itself as it keeps
that interface simpler, and keeps the true nature of the socket available
to tcpmon if needed.