Added tests for table manager and fixed bg manager tests.
Table manager now has tests to ensure claim and heir functionality.
The background manager tests adhere to the usual scheme for test module
naming.

Version of rebar used uses 'suites' rather than 'suite' to specific
which suites of tests to run.

The table manager will allow up to 2 processes to claim a table in an
effort to avoid a race contition. Consider a process that is supervised
exiting and being restarted by a supervisor. If the process is restarted
and tries to claim the table before the ets-transfer message has reached
the table manager, it may be incorrectly denied the claim. By allowing it
to register and blocking it until the claim can be verified (ie the table
manager gets the ets transfer message), this condition is avoided.