Fix hashtree:sha_test_ from timing out
EQC takes a long time, and uses gigabytes of memory to generate 1MB
random binaries, as evidenced by:

    eqc_gen:sample(eqc_gen:binary(1024 * 1024)).

We're still able to write a solid test, by ensuring that we always chunk
the binary into between 1 and 16 chunks. So instead of generating the
chunk size, we now generate the number of chunks, and derive chunk size
from that. This change allowed us to remove one level of ?FORALL
nesting, as well.